// packages/ticket-thread/src/voice-calls/models/voiceCallEventTypes.ts
export enum PhoneIntegrationEvent {
    ConversationStarted = 'phone-call-conversation-started',
    PhoneCallForwardedToExternalNumber = 'phone-call-forwarded-to-external-number',
    PhoneCallForwardedToGorgiasNumber = 'phone-call-forwarded-to-gorgias-number',
    PhoneCallForwarded = 'phone-call-forwarded',
    MessagePlayed = 'message-played',
    DeclinedPhoneCall = 'declined-phone-call',
    OutgoingPhoneCallConnected = 'outgoing-phone-call-connected',
    PhoneCallRinging = 'phone-call-ringing',
    ChildCallNotAnswered = 'child-call-not-answered',
    IncomingPhoneCall = 'incoming-phone-call',
    OutgoingPhoneCall = 'outgoing-phone-call',
    CompletedPhoneCall = 'completed-phone-call',
    CallRecording = 'call-recording',
    MissedPhoneCall = 'missed-phone-call',
    VoicemailRecording = 'voicemail-recording',
    PhoneCallAnswered = 'phone-call-answered',
    PhoneCallTransferInitiated = 'phone-call-transfer-initiated',
    PhoneCallTransferFailed = 'phone-call-transfer-failed',
    EndingTriggered = 'call-ending-triggered',
    Enqueued = 'phone-call-enqueued',
    Dequeued = 'phone-call-dequeued',
    IvrOptionSelected = 'ivr-option-selected',
}